Q1 almost done
Changed it so its not longer “Score + -#” but rather “Score - #”.  I
find this easier to read.  Something may be wrong with the distance
methods, as setting the default to 99999 was causing the code to hang a
bit.

Currently, we are getting 9/10 of the test cases right on the auto
grader, and getting 3/4 points.